Feature/auth policy/8

+++ app/Http/Controllers/CommentController.php
@@ -8,9 +8,18 @@ use App\Models\Comment;
 use App\Models\Theme;
 use App\Services\CommentControllerService;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Log;
 
 class CommentController extends Controller
 {
+    protected $user;
+
+    public function __construct() {
+        $this->user = Auth::user();
+    }
+
     /*
     *  Отфильтровать комментариии по темам
     */
@@ -29,7 +38,8 @@ class CommentController extends Controller
     */
     public function left(Theme $topic, CommentSendRequest $request, CommentControllerService $service): array
     {
-        $comment = $service->createFromRequest($request->validated(), $topic);
+        Gate::authorize('create', Comment::class);
+        $comment = $service->createFromRequest($request->validated(), $topic, $this->user);
         return [
             'status' => 'success',
             'comment' => $comment,
@@ -41,6 +51,7 @@ class CommentController extends Controller
     */
     public function update(Comment $comment, CommentUpdateRequest $request, CommentControllerService $service): array
     {
+        Gate::authorize('update', $comment);
         $comment = $service->updateFromRequest($comment, $request->validated());
         return [
             'status' => 'success',
@@ -53,6 +64,7 @@ class CommentController extends Controller
     */
     public function delete(Comment $comment, CommentControllerService $service): array
     {
+        Gate::authorize('delete', $comment);
         $service->delete($comment);
         return [
             'status' => 'success',
